Be consistent about (size_t)(-1)
    $ git grep "nkeys" | grep "\-1"
produced 4/6 instances of "(size_t)(-1)" and 2/6 instances of
"-1".  Since storing -1 in a (size_t) is slightly unintuitive (for
inexpert C programmers), I think it makes sense to standardize on
being explicit.

+++ kvlds/btree_node_merge.c
@@ -138,7 +138,7 @@ merge_parent(struct btree * T, struct node ** c_in,
 	for (i = 0; i <= nsep; i++) {
 		free(c_in[i]->u.keys);
 		free(c_in[i]->v.children);
-		c_in[i]->nkeys = -1;
+		c_in[i]->nkeys = (size_t)(-1);
 		btree_node_destroy(T, c_in[i]);
 	}
